+++ src/RulesEngine/ExpressionBuilders/LambdaExpressionBuilder.cs
@@ -4,12 +4,12 @@
 using RulesEngine.HelperFunctions;
 using RulesEngine.Models;
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
 
 namespace RulesEngine.ExpressionBuilders
 {
-    /// <summary>
-    /// This class will build the list expression
-    /// </summary>
     internal sealed class LambdaExpressionBuilder : RuleExpressionBuilderBase
     {
         private readonly ReSettings _reSettings;
@@ -26,19 +26,32 @@ namespace RulesEngine.ExpressionBuilders
             try
             {
                 var ruleDelegate = _ruleExpressionParser.Compile<bool>(rule.Expression, ruleParams);
-                bool func(object[] paramList) => ruleDelegate(paramList);
-                return Helpers.ToResultTree(rule, null, func);
+                return Helpers.ToResultTree(rule, null, ruleDelegate);
             }
             catch (Exception ex)
             {
                 ex.Data.Add(nameof(rule.RuleName), rule.RuleName);
                 ex.Data.Add(nameof(rule.Expression), rule.Expression);
 
-                if (!_reSettings.EnableExceptionAsErrorMessage) throw;
+                if (!_reSettings.EnableExceptionAsErrorMessage)
+                {
+                    throw;
+                }
+
                 bool func(object[] param) => false;
-                var exceptionMessage = $"Exception while parsing expression `{rule?.Expression}` - {ex.Message}";
-                return Helpers.ToResultTree(rule, null, func, exceptionMessage);
+                var exceptionMessage = _reSettings.IgnoreException ? "" : $"Exception while parsing expression `{rule?.Expression}` - {ex.Message}";
+                return Helpers.ToResultTree(rule, null,func, exceptionMessage);
             }
+        }
+
+        internal override LambdaExpression Parse(string expression, ParameterExpression[] parameters, Type returnType)
+        {
+            return _ruleExpressionParser.Parse(expression, parameters, returnType);
+        }
+
+        internal override Func<object[],Dictionary<string,object>> CompileScopedParams(RuleParameter[] ruleParameters, RuleExpressionParameter[] scopedParameters)
+        {
+            return _ruleExpressionParser.CompileRuleExpressionParameters(ruleParameters, scopedParameters);
         }
     }
 }