+++ src/RulesEngine/ExpressionBuilders/RuleExpressionBuilderBase.cs
@@ -2,6 +2,9 @@
 // Licensed under the MIT License.
 
 using RulesEngine.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq.Expressions;
 
 namespace RulesEngine.ExpressionBuilders
 {
@@ -18,5 +21,9 @@ namespace RulesEngine.ExpressionBuilders
         /// <param name="ruleInputExp">The rule input exp.</param>
         /// <returns>Expression type</returns>
         internal abstract RuleFunc<RuleResultTree> BuildDelegateForRule(Rule rule, RuleParameter[] ruleParams);
+
+        internal abstract LambdaExpression Parse(string expression, ParameterExpression[] parameters, Type returnType);
+
+        internal abstract Func<object[], Dictionary<string, object>> CompileScopedParams(RuleParameter[] ruleParameters, RuleExpressionParameter[] scopedParameters);
     }
 }