+++ src/RulesEngine/HelperFunctions/Helpers.cs
@@ -29,6 +29,7 @@ namespace RulesEngine.HelperFunctions
         /// </summary>
         /// <param name="ruleResultTree">ruleResultTree</param>
         /// <param name="ruleResultMessage">ruleResultMessage</param>
+        [Obsolete]
         internal static void ToResultTreeMessages(RuleResultTree ruleResultTree, ref RuleResultMessage ruleResultMessage)
         {
             if (ruleResultTree.ChildResults != null)
@@ -40,7 +41,7 @@ namespace RulesEngine.HelperFunctions
                 if (!ruleResultTree.IsSuccess)
                 {
                     string errMsg = ruleResultTree.Rule.ErrorMessage;
-                    errMsg = string.IsNullOrEmpty(errMsg) ? $"Error message does not configured for {ruleResultTree.Rule.RuleName}" : errMsg;
+                    errMsg = string.IsNullOrEmpty(errMsg) ? $"Error message is not configured for {ruleResultTree.Rule.RuleName}" : errMsg;
 
                     if (ruleResultTree.Rule.ErrorType == ErrorType.Error && !ruleResultMessage.ErrorMessages.Contains(errMsg))
                     {
@@ -59,6 +60,7 @@ namespace RulesEngine.HelperFunctions
         /// </summary>
         /// <param name="childResultTree">childResultTree</param>
         /// <param name="ruleResultMessage">ruleResultMessage</param>
+        [Obsolete]
         private static void GetChildRuleMessages(IEnumerable<RuleResultTree> childResultTree, ref RuleResultMessage ruleResultMessage)
         {
             foreach (var item in childResultTree)