+++ src/RulesEngine/Models/ScopedParam.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation.
 //  Licensed under the MIT License.
 
-using Newtonsoft.Json;
 using System.Diagnostics.CodeAnalysis;
 
 namespace RulesEngine.Models
@@ -9,22 +8,23 @@ namespace RulesEngine.Models
     /// <summary>Class LocalParam.
     /// </summary>
     [ExcludeFromCodeCoverage]
-    public class LocalParam
+    public class ScopedParam
     {
 
         /// <summary>
-        /// Gets or sets the name of the rule.
+        /// Gets or sets the name of the param.
         /// </summary>
         /// <value>
         /// The name of the rule.
-        /// </value>
-        [JsonProperty, JsonRequired]
+        /// </value>]
         public string Name { get; set; }
 
         /// <summary>
-        /// Gets or Sets the lambda expression. 
+        /// Gets or Sets the lambda expression which can be reference in Rule. 
         /// </summary>
-        [JsonProperty, JsonRequired]
         public string Expression { get; set; }
     }
+
+    [ExcludeFromCodeCoverage]
+    public class LocalParam : ScopedParam { }
 }