+++ src/RulesEngine/RuleCompiler.cs
@@ -2,12 +2,14 @@
 // Licensed under the MIT License.
 
 using Microsoft.Extensions.Logging;
+using RulesEngine.ExpressionBuilders;
 using RulesEngine.HelperFunctions;
 using RulesEngine.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
+using System.Runtime.InteropServices;
 
 namespace RulesEngine
 {
@@ -25,6 +27,7 @@ namespace RulesEngine
         /// The expression builder factory
         /// </summary>
         private readonly RuleExpressionBuilderFactory _expressionBuilderFactory;
+        private readonly ReSettings _reSettings;
 
         /// <summary>
         /// The logger
@@ -36,10 +39,12 @@ namespace RulesEngine
         /// </summary>
         /// <param name="expressionBuilderFactory">The expression builder factory.</param>
         /// <exception cref="ArgumentNullException">expressionBuilderFactory</exception>
-        internal RuleCompiler(RuleExpressionBuilderFactory expressionBuilderFactory, ILogger logger)
+        internal RuleCompiler(RuleExpressionBuilderFactory expressionBuilderFactory, ReSettings reSettings, ILogger logger)
         {
             _logger = logger ?? throw new ArgumentNullException($"{nameof(logger)} can't be null.");
+          
             _expressionBuilderFactory = expressionBuilderFactory ?? throw new ArgumentNullException($"{nameof(expressionBuilderFactory)} can't be null.");
+            _reSettings = reSettings;
         }
 
         /// <summary>
@@ -50,7 +55,7 @@ namespace RulesEngine
         /// <param name="input"></param>
         /// <param name="ruleParam"></param>
         /// <returns>Compiled func delegate</returns>
-        internal RuleFunc<RuleResultTree> CompileRule(Rule rule, params RuleParameter[] ruleParams)
+        internal RuleFunc<RuleResultTree> CompileRule(Rule rule, RuleParameter[] ruleParams, ScopedParam[] globalParams)
         {
             try
             {
@@ -58,8 +63,11 @@ namespace RulesEngine
                 {
                     throw new ArgumentNullException(nameof(rule));
                 }
-                var ruleExpression = GetDelegateForRule(rule, ruleParams);
-                return ruleExpression;
+                var globalParamExp = GetRuleExpressionParameters(rule.RuleExpressionType,globalParams, ruleParams);
+                var extendedRuleParams = ruleParams.Concat(globalParamExp.Select(c => new RuleParameter(c.ParameterExpression.Name,c.ParameterExpression.Type)))
+                                                   .ToArray();
+                var ruleExpression = GetDelegateForRule(rule, extendedRuleParams);
+                return GetWrappedRuleFunc(RuleExpressionType.LambdaExpression,ruleExpression,ruleParams,globalParamExp);
             }
             catch (Exception ex)
             {
@@ -79,15 +87,55 @@ namespace RulesEngine
         /// <returns></returns>
         private RuleFunc<RuleResultTree> GetDelegateForRule(Rule rule, RuleParameter[] ruleParams)
         {
+            var scopedParamList = GetRuleExpressionParameters(rule.RuleExpressionType, rule?.LocalParams, ruleParams);
+
+            var extendedRuleParams = ruleParams.Concat(scopedParamList.Select(c => new RuleParameter(c.ParameterExpression.Name, c.ParameterExpression.Type)))
+                                               .ToArray();
+
+            RuleFunc<RuleResultTree> ruleFn;
+            
             if (Enum.TryParse(rule.Operator, out ExpressionType nestedOperator) && nestedOperators.Contains(nestedOperator) &&
                 rule.Rules != null && rule.Rules.Any())
             {
-                return BuildNestedRuleFunc(rule, nestedOperator, ruleParams);
+                ruleFn = BuildNestedRuleFunc(rule, nestedOperator, extendedRuleParams);
             }
             else
             {
-                return BuildRuleFunc(rule, ruleParams);
+                ruleFn = BuildRuleFunc(rule, extendedRuleParams);
             }
+
+            return GetWrappedRuleFunc(rule.RuleExpressionType, ruleFn, ruleParams, scopedParamList);
+        }
+
+        private RuleExpressionParameter[] GetRuleExpressionParameters(RuleExpressionType ruleExpressionType,IEnumerable<ScopedParam> localParams, RuleParameter[] ruleParams)
+        {
+            if(!_reSettings.EnableScopedParams)
+            {
+                return new RuleExpressionParameter[] { };
+            }
+            var ruleExpParams = new List<RuleExpressionParameter>();
+
+            if (localParams?.Any() == true)
+            {
+
+                var parameters = ruleParams.Select(c => c.ParameterExpression)
+                                            .ToList();
+
+                var expressionBuilder = GetExpressionBuilder(ruleExpressionType);
+
+                foreach (var lp in localParams)
+                {
+                    var lpExpression = expressionBuilder.Parse(lp.Expression, parameters.ToArray(), null).Body;
+                    var ruleExpParam = new RuleExpressionParameter() {
+                        ParameterExpression = Expression.Parameter(lpExpression.Type, lp.Name),
+                        ValueExpression = lpExpression
+                    };
+                    parameters.Add(ruleExpParam.ParameterExpression);
+                    ruleExpParams.Add(ruleExpParam);
+                }
+            }
+            return ruleExpParams.ToArray();
+
         }
 
         /// <summary>
@@ -100,12 +148,7 @@ namespace RulesEngine
         /// <exception cref="InvalidOperationException"></exception>
         private RuleFunc<RuleResultTree> BuildRuleFunc(Rule rule, RuleParameter[] ruleParams)
         {
-            if (!rule.RuleExpressionType.HasValue)
-            {
-                throw new InvalidOperationException($"RuleExpressionType can not be null for leaf level expressions.");
-            }
-
-            var ruleExpressionBuilder = _expressionBuilderFactory.RuleGetExpressionBuilder(rule.RuleExpressionType.Value);
+            var ruleExpressionBuilder = GetExpressionBuilder(rule.RuleExpressionType);
 
             var ruleFunc = ruleExpressionBuilder.BuildDelegateForRule(rule, ruleParams);
 
@@ -125,13 +168,13 @@ namespace RulesEngine
         private RuleFunc<RuleResultTree> BuildNestedRuleFunc(Rule parentRule, ExpressionType operation, RuleParameter[] ruleParams)
         {
             var ruleFuncList = new List<RuleFunc<RuleResultTree>>();
-            foreach (var r in parentRule.Rules)
+            foreach (var r in parentRule.Rules.Where(c => c.Enabled))
             {
                 ruleFuncList.Add(GetDelegateForRule(r, ruleParams));
             }
 
             return (paramArray) => {
-                var resultList = ruleFuncList.Select(fn => fn(paramArray));
+                var resultList = ruleFuncList.Select(fn => fn(paramArray)).ToList();
                 Func<object[], bool> isSuccess = (p) => ApplyOperation(resultList, operation);
                 var result = Helpers.ToResultTree(parentRule, resultList, isSuccess);
                 return result(paramArray);
@@ -141,6 +184,11 @@ namespace RulesEngine
 
         private bool ApplyOperation(IEnumerable<RuleResultTree> ruleResults, ExpressionType operation)
         {
+            if (ruleResults?.Any() != true)
+            {
+                return false;
+            }
+
             switch (operation)
             {
                 case ExpressionType.And:
@@ -153,6 +201,37 @@ namespace RulesEngine
                 default:
                     return false;
             }
+        }
+
+        private RuleFunc<RuleResultTree> GetWrappedRuleFunc(RuleExpressionType ruleExpressionType, RuleFunc<RuleResultTree> ruleFunc,RuleParameter[] ruleParameters,RuleExpressionParameter[] ruleExpParams)
+        {
+            if(ruleExpParams.Length == 0)
+            {
+                return ruleFunc;
+            }
+            var paramDelegate = GetExpressionBuilder(ruleExpressionType).CompileScopedParams(ruleParameters, ruleExpParams);
+
+            return (ruleParams) => {
+                var inputs = ruleParams.Select(c => c.Value).ToArray();
+                var scopedParamsDict = paramDelegate(inputs);
+                var scopedParams = scopedParamsDict.Select(c => new RuleParameter(c.Key, c.Value)).ToList();
+                var extendedInputs = ruleParams.Concat(scopedParams);
+                var result = ruleFunc(extendedInputs.ToArray());
+                // To be removed in next major release
+#pragma warning disable CS0618 // Type or member is obsolete
+                if(result.RuleEvaluatedParams == null)
+                {
+                    result.RuleEvaluatedParams = scopedParams;
+                }
+#pragma warning restore CS0618 // Type or member is obsolete
+
+                return result;
+            };
+        }
+
+        private RuleExpressionBuilderBase GetExpressionBuilder(RuleExpressionType expressionType)
+        {
+            return _expressionBuilderFactory.RuleGetExpressionBuilder(expressionType);
         }
     }
 }