+++ src/RulesEngine/RulesCache.cs
@@ -27,6 +27,11 @@ namespace RulesEngine
             return _workflowRules.ContainsKey(workflowName);
         }
 
+        public List<string> GetAllWorkflowNames()
+        {
+            return _workflowRules.Keys.ToList();
+        }
+
         /// <summary>Determines whether [contains compiled rules] [the specified workflow name].</summary>
         /// <param name="workflowName">Name of the workflow.</param>
         /// <returns>
@@ -57,16 +62,6 @@ namespace RulesEngine
         {
             _workflowRules.Clear();
             _compileRules.Clear();
-        }
-
-        /// <summary>Gets the rules.</summary>
-        /// <param name="workflowName">Name of the workflow.</param>
-        /// <returns>IEnumerable&lt;Rule&gt;.</returns>
-        public IEnumerable<Rule> GetRules(string workflowName)
-        {
-            if (!ContainsWorkflowRules(workflowName))
-                throw new ArgumentException($"workflow `{workflowName}` was not found");
-            return _workflowRules[workflowName].Rules;
         }
 
         /// <summary>Gets the work flow rules.</summary>