+++ src/RulesEngine/RulesEngine.cs
@@ -32,8 +32,6 @@ namespace RulesEngine
         private readonly ILogger _logger;
         private readonly ReSettings _reSettings;
         private readonly RulesCache _rulesCache = new RulesCache();
-        private MemoryCache _compiledParamsCache = new MemoryCache(new MemoryCacheOptions());
-        private readonly ParamCompiler _ruleParamCompiler;
         private readonly RuleExpressionParser _ruleExpressionParser;
         private readonly RuleCompiler _ruleCompiler;
         private readonly ActionFactory _actionFactory;
@@ -57,8 +55,7 @@ namespace RulesEngine
             _logger = logger ?? new NullLogger<RulesEngine>();
             _reSettings = reSettings ?? new ReSettings();
             _ruleExpressionParser = new RuleExpressionParser(_reSettings);
-            _ruleParamCompiler = new ParamCompiler(_reSettings, _ruleExpressionParser);
-            _ruleCompiler = new RuleCompiler(new RuleExpressionBuilderFactory(_reSettings, _ruleExpressionParser), _logger);
+            _ruleCompiler = new RuleCompiler(new RuleExpressionBuilderFactory(_reSettings, _ruleExpressionParser),_reSettings, _logger);
             _actionFactory = new ActionFactory(GetActionRegistry(_reSettings));
         }
 
@@ -173,13 +170,17 @@ namespace RulesEngine
             }
         }
 
+        public List<string> GetAllRegisteredWorkflowNames()
+        {
+            return _rulesCache.GetAllWorkflowNames();
+        }
+
         /// <summary>
         /// Clears the workflows.
         /// </summary>
         public void ClearWorkflows()
         {
             _rulesCache.Clear();
-            ClearCompiledParamCache();
         }
 
         /// <summary>
@@ -192,7 +193,6 @@ namespace RulesEngine
             {
                 _rulesCache.Remove(workflowName);
             }
-            ClearCompiledParamCache();
         }
 
         /// <summary>
@@ -239,9 +239,9 @@ namespace RulesEngine
             if (workflowRules != null)
             {
                 var dictFunc = new Dictionary<string, RuleFunc<RuleResultTree>>();
-                foreach (var rule in workflowRules.Rules)
+                foreach (var rule in workflowRules.Rules.Where(c => c.Enabled))
                 {
-                    dictFunc.Add(rule.RuleName, CompileRule(workflowName, ruleParams, rule));
+                    dictFunc.Add(rule.RuleName, CompileRule(rule, ruleParams, workflowRules.GlobalParams?.ToArray()));
                 }
 
                 _rulesCache.AddOrUpdateCompiledRule(compileRulesKey, dictFunc);
@@ -257,42 +257,22 @@ namespace RulesEngine
 
         private RuleFunc<RuleResultTree> CompileRule(string workflowName, string ruleName, RuleParameter[] ruleParameters)
         {
-            var rules = _rulesCache.GetRules(workflowName);
-            var currentRule = rules?.SingleOrDefault(c => c.RuleName == ruleName);
+            var workflow = _rulesCache.GetWorkFlowRules(workflowName);
+            if(workflow == null)
+            {
+                throw new ArgumentException($"Workflow `{workflowName}` is not found");
+            }
+            var currentRule = workflow.Rules?.SingleOrDefault(c => c.RuleName == ruleName && c.Enabled);
             if (currentRule == null)
             {
                 throw new ArgumentException($"Workflow `{workflowName}` does not contain any rule named `{ruleName}`");
             }
-            return CompileRule(workflowName, ruleParameters, currentRule);
+            return CompileRule(currentRule, ruleParameters, workflow.GlobalParams?.ToArray());
         }
 
-        private RuleFunc<RuleResultTree> CompileRule(string workflowName, RuleParameter[] ruleParams, Rule rule)
+        private RuleFunc<RuleResultTree> CompileRule(Rule rule, RuleParameter[] ruleParams, ScopedParam[] scopedParams)
         {
-            if (!_reSettings.EnableLocalParams)
-            {
-                return _ruleCompiler.CompileRule(rule, ruleParams);
-            }
-            var compiledParamsKey = GetCompiledParamsCacheKey(workflowName, rule.RuleName, ruleParams);
-            var compiledParamList = _compiledParamsCache.GetOrCreate(compiledParamsKey, (entry) => _ruleParamCompiler.CompileParamsExpression(rule, ruleParams));
-            var compiledRuleParameters = compiledParamList?.Select(c => c.AsRuleParameter()) ?? new List<RuleParameter>();
-            var updatedRuleParams = ruleParams?.Concat(compiledRuleParameters);
-            var compiledRule = _ruleCompiler.CompileRule(rule, updatedRuleParams?.ToArray());
-
-            RuleFunc<RuleResultTree> updatedRule = (RuleParameter[] paramList) => {
-                var inputs = paramList.AsEnumerable();
-                var localParams = compiledParamList ?? new List<CompiledParam>();
-                var evaluatedParamList = new List<RuleParameter>();
-                foreach (var localParam in localParams)
-                {
-                    var evaluatedLocalParam = _ruleParamCompiler.EvaluateCompiledParam(localParam.Name, localParam.Value, inputs);
-                    inputs = inputs.Append(evaluatedLocalParam);
-                    evaluatedParamList.Add(evaluatedLocalParam);
-                }
-                var result = compiledRule(inputs.ToArray());
-                result.RuleEvaluatedParams = evaluatedParamList;
-                return result;
-            };
-            return updatedRule;
+            return _ruleCompiler.CompileRule(rule, ruleParams, scopedParams);
         }
 
 
@@ -325,12 +305,6 @@ namespace RulesEngine
             return key;
         }
 
-        private string GetCompiledParamsCacheKey(string workflowName, string ruleName, RuleParameter[] ruleParams)
-        {
-            var key = $"compiledparams-{workflowName}-{ruleName}" + string.Join("-", ruleParams.Select(c => c.Type.Name));
-            return key;
-        }
-
         private IDictionary<string, Func<ActionBase>> GetDefaultActionRegistry()
         {
             return new Dictionary<string, Func<ActionBase>>{
@@ -351,7 +325,7 @@ namespace RulesEngine
                 foreach (var ruleResult in ruleResultList?.Where(r => !r.IsSuccess))
                 {
                     var errorMessage = ruleResult?.Rule?.ErrorMessage;
-                    if (errorMessage != null)
+                    if (string.IsNullOrWhiteSpace(ruleResult.ExceptionMessage) && errorMessage != null)
                     {
                         var errorParameters = Regex.Matches(errorMessage, ParamParseRegex);
 
@@ -406,16 +380,6 @@ namespace RulesEngine
 
             return errorMessage;
         }
-
-        /// <summary>
-        /// Clears all compiledParams
-        /// </summary>
-        private void ClearCompiledParamCache()
-        {
-            _compiledParamsCache.Dispose();
-            _compiledParamsCache = new MemoryCache(new MemoryCacheOptions());
-        }
-
         #endregion
     }
 }