+++ src/RulesEngine/Validators/RuleValidator.cs
@@ -19,7 +19,7 @@ namespace RulesEngine.Validators
             RuleFor(c => c.RuleName).NotEmpty().WithMessage(Constants.RULE_NAME_NULL_ERRMSG);
 
             //Nested expression check
-            When(c => c.RuleExpressionType == null, () => {
+            When(c => c.Operator != null, () => {
                 RuleFor(c => c.Operator)
                    .NotNull().WithMessage(Constants.OPERATOR_NULL_ERRMSG)
                    .Must(op => _nestedOperators.Any(x => x.ToString().Equals(op, StringComparison.OrdinalIgnoreCase)))
@@ -37,9 +37,8 @@ namespace RulesEngine.Validators
 
         private void RegisterExpressionTypeRules()
         {
-            When(c => c.RuleExpressionType == RuleExpressionType.LambdaExpression, () => {
+            When(c => c.Operator == null && c.RuleExpressionType == RuleExpressionType.LambdaExpression, () => {
                 RuleFor(c => c.Expression).NotEmpty().WithMessage(Constants.LAMBDA_EXPRESSION_EXPRESSION_NULL_ERRMSG);
-                RuleFor(c => c.Operator).Null().WithMessage(Constants.LAMBDA_EXPRESSION_OPERATOR_ERRMSG);
                 RuleFor(c => c.Rules).Null().WithMessage(Constants.LAMBDA_EXPRESSION_RULES_ERRMSG);
             });
         }