+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [3.1.0-preview.1]
+- Added globalParams feature which can be applied to all rules
+- Enabled localParams support for nested Rules
+- Made certain fields in Rule model optional allowing users to define workflow with minimal fields
+- Added option to disable Rule in workflow json
+- Added `GetAllRegisteredWorkflow` to RulesEngine to return all registeredWorkflows
+- Fixed Rule compilation exception not returned when Rule has ErrorMessage field defined - #95
+
 ## [3.0.2]
 - Fixed LocalParams cache not getting cleaned up when RemoveWorkflow and ClearWorkflows are called